fixed hw8
