+++ src/main/java/busTickets/BusTicketClass.java
@@ -1,4 +1,4 @@
-package busTicketHandler;
+package busTickets;
 
 public enum BusTicketClass {
     CLA, STD
