+++ src/main/java/busTickets/BusTicketType.java
@@ -1,4 +1,4 @@
-package busTicketHandler;
+package busTickets;
 
 public enum BusTicketType {
     DAY,WEEK,MONTH,YEAR,PRIME
